+++ TabletopTweaks/NewContent/ContentAdder.cs
@@ -56,6 +56,7 @@ namespace TabletopTweaks.NewContent {
 
                 WizardArcaneDiscoveries.ArcaneDiscoverySelection.AddArcaneDiscoverySelection();
                 WizardArcaneDiscoveries.OppositionResearch.AddOppositionResearch();
+                WizardArcaneDiscoveries.KnowledgeIsPower.AddKnowledgeIsPower();
 
                 Features.AeonBaneIncreaseResourceFeature.AddAeonBaneIncreaseResourceFeature();
                 Features.PrimalistRagePowerSelection.AddPrimalistRagePowerSelection();